frontend/requirements.txt
pillow>=10.0.0,<11.0.0
gradio==4.19.2
huggingface-hub<0.26.0
python-dotenv==1.0.1
pandas 
plotly
opencv-python